rag_backend/requirements.txt
Flask
PyMuPDF
SQLAlchemy
beautifulsoup4
faiss-cpu
fastapi
gunicorn
langchain
langchain-community
langchain-openai
lxml
openai
openai>=1.0.0
pgvector
pgvector
pgvector 
psycopg2-binary
psycopg2-binary
python-dotenv
python-dotenv
requests
sentence-transformers
tiktoken
typer
unstructured[md]
uvicorn[standard]
